Join dqa3_template on tr.set_key rather than data.set_key
This was to work around an issue with the DQA3 database where what
appears to be some sort of race condition in the DQA3 software due to
two records on different machines having identical measured times.  This
resulted in the dqa3_trend.data_key being set to an invalid, duplicated
value.  The net effect was that QCPump sent data to the wrong unit.

Since the set_key on the dqa3_trend table still points to the correct
template row, joining the template table with the dqa3_trend table
instead of the dqa3_data table results in the data being sent to
the correct unit.  Unfortunately in this case the temparture, pressure,
comment, signature, and device serial number will be incorrect since
they are pulled from the dqa3_data table.

More info available in private communication between Randy Taylor &
Keith Offer.